+++ src/main/java/com/grepp/spring/app/controller/api/ChatController.java
@@ -1,9 +1,8 @@
 package com.grepp.spring.app.controller.api;
 
-import com.grepp.spring.app.controller.api.chat.ChatCursorRequest;
-import com.grepp.spring.app.controller.api.chat.ChatHistoryResponse;
-import com.grepp.spring.app.controller.api.chat.ChatPageResponse;
-import com.grepp.spring.app.controller.api.chat.ParticipantResponse;
+import com.grepp.spring.app.model.chat.dto.request.ChatCursorRequest;
+import com.grepp.spring.app.model.chat.dto.response.ChatPageResponse;
+import com.grepp.spring.app.model.chat.dto.response.ParticipantResponse;
 import com.grepp.spring.app.model.auth.domain.Principal;
 import com.grepp.spring.app.model.chat.service.ChatService;
 import com.grepp.spring.app.model.study.service.StudyService;
@@ -14,7 +13,6 @@ import com.grepp.spring.infra.util.SecurityUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -26,8 +24,6 @@ import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
