+++ src/main/java/com/grepp/spring/app/model/chat/dto/request/ChatCursorRequest.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.controller.api.chat;
+package com.grepp.spring.app.model.chat.dto.request;
 
 import java.time.LocalDateTime;
 import org.springframework.format.annotation.DateTimeFormat;
