+++ src/main/java/com/grepp/spring/app/model/chat/dto/response/SessionUserInfo.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.controller.api.chat;
+package com.grepp.spring.app.model.chat.dto.response;
 
 public record SessionUserInfo (
     Long memberId,
