+++ src/main/java/com/grepp/spring/infra/config/Chat/WebSocket/WebSocketHandler.java
@@ -1,6 +1,6 @@
 package com.grepp.spring.infra.config.Chat.WebSocket;
 
-import com.grepp.spring.app.controller.api.chat.SessionUserInfo;
+import com.grepp.spring.app.model.chat.dto.response.SessionUserInfo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.WebSocketSession;
